chore: remove redundant client directories
